requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
bs4==0.0.1
celery==4.1.0
dj-database-url==0.4.2
Django==2.0
django-celery-results==1.0.1
django-cors-headers==2.1.0
django-filter==0.12.0
django-templated-mail==1.0.0
djangorestframework==3.7.3
djoser==1.1.4
gunicorn==19.7.1
kombu==4.1.0
Markdown==2.6.5
pytz==2017.2
redis==2.10.6
requests==2.10.0
vine==1.1.4
whitenoise==2.0.6
